refactor: move self destructs calculation to typescript
